api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==3.1.7",
    "fastapi[standard]>=0.128.8",
    "langchain-community[standard]>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai[standard]>=1.1.9",
    "langchain[standard]>=1.2.10",
    "passlib>=1.7.4",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings[standard]>=2.12.0",
    "pypdf>=6.7.1",
    "python-jose[cryptography]>=3.5.0",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.3",
    "tavily-python>=0.7.21",
    "uvicorn[standard]>=0.40.0",
]
